backend/requirements.txt
dj-database-url>=2.1.0
django-filter>=24.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
pillow>=10.3.0
PyJWT>=2.8.0
pytz>=2024.1
whitenoise>=6.6.0
asgiref>=3.7.2
Django>=3.2.19
gunicorn>=20.1.0
packaging>=24.0
sqlparse>=0.4.4
typing_extensions>=4.6.3
psycopg2-binary>=2.9.6
django-cors-headers>=3.10.0
